requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ffprobe==0.5
flatbuffers==24.3.25
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
humanfriendly==10.0
idna==3.7
imageio==2.34.2
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.2
pyarrow==16.1.0
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.0
Pygments==2.18.0
PyMatting==1.1.12
pyshortners==1.0.22
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
rembg==2.0.57
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.24.0
scipy==1.14.0
setuptools==70.2.0
shortzyy==0.0.6
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.36.0
streamlit-chat==0.1.1
streamlit-lottie==0.0.5
sympy==1.12.1
tenacity==8.4.2
tifffile==2024.7.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wheel==0.43.0
yarl==1.9.4
